chore: rename to runner.rs and update architecture.md
