+++ crates/rover-client/src/operations/config/who_am_i/mod.rs
@@ -1,4 +1,4 @@
 mod types;
 
-pub mod query_runner;
+pub mod runner;
 pub use types::{Actor, ConfigWhoAmIInput, RegistryIdentity};
