+++ crates/rover-client/src/operations/graph/introspect/mod.rs
@@ -1,7 +1,7 @@
-mod query_runner;
+mod runner;
 mod schema;
 mod types;
 
-pub use query_runner::run;
+pub use runner::run;
 pub use schema::Schema;
 pub use types::{GraphIntrospectInput, GraphIntrospectResponse};
