+++ crates/rover-client/src/operations/subgraph/delete/runner.rs
