+++ crates/rover-client/src/operations/subgraph/introspect/mod.rs
@@ -1,5 +1,5 @@
-pub(crate) mod query_runner;
+pub(crate) mod runner;
 pub(crate) mod types;
 
-pub use query_runner::run;
+pub use runner::run;
 pub use types::{SubgraphIntrospectInput, SubgraphIntrospectResponse};
