+++ crates/rover-client/src/operations/subgraph/publish/mod.rs
@@ -1,3 +1,3 @@
-pub mod mutation_runner;
+pub mod runner;
 pub(crate) mod types;
 pub use types::{SubgraphPublishInput, SubgraphPublishResponse};
